+++ riscv/insns/dlxfer_u.h
@@ -5,8 +5,8 @@ if (proxy == nullptr) return -1;
 auto dst = RD - 0xf8000000;
 auto src = *((unsigned int *)MMU.get_phy_addr(RS1));
 
-unsigned short col = DMA_SHAPE_COLUMN;
-unsigned short row = DMA_SHAPE_ROW;
+unsigned int col = DMA_SHAPE_COLUMN;
+unsigned int row = DMA_SHAPE_ROW;
 unsigned int stride = STRIDE_DDR;
 
 for (int times = 0; times < 5; times++) {
